[YOURLSClient] Cog model restructure (#624)
* [YOURLSClient] Cog model restructure

Restructure cog `YOURLSClient` according to SFUAnime/Ren#608. No
functional changes. Constants are moved to `constants.py`.

* [YOURLSClient] Use logging date format YYYY-MM-DD

Use logging date format YYYY-MM-DD instead of DD-MM-YYYY.

* [YOURLSClient] Add docstring to `YOURLS` class

Instead of doing `pass`, a docstring is attached to `YOURLS` class.
It is used in the help string of the cog name if we run the `[p]help
yourls` command.