[Heartbeat] Cog model restructure (#627)
* [Heartbeat] Cog model restructure

Restructure cog `Heartbeat` according to SFUAnime/Ren#608. No functional
changes.

Constants are moved to `constants.py`.

* [Heartbeat] Add unit tests for `CommandsCore`

This commit adds unit tests for all methods in `CommandsCore`.

* [Heartbeat] Add unit tests for `Core`

This commit adds unit tests for the main loop (i.e., method `_loop`) in
class `Core`.

* [Heartbeat] Add pytest fixtures

Add relevant pytest fixtures to be used for refactoring
`testCommandsCore` and `testCore`.

* [Heartbeat] Refactor `testCommandsCore.py`

- Use avaliable pytest fixtures.
- Add function `contentFromMockContextSend` to contain the logic
  obtaining `content of mock `Context.send`.

* [Heartbeat] Refactor `testCore.py`

- Use available pytest fixtures `event_loop` and `cogHeartbeat`.
- Remove `red.remove_cog()` call because it is not a part of `_loop`
  test.

* [Heartbeat] Mock out `asyncio.sleep` in `testCore`

Since we do not care about whether the real time sleep duration is as
expected, but rather the fact that the loop has retried a certain number
of times as expected, and that the loop fails after receiving bad
respones, thus, `asyncio.sleep` can be mock-patched so that we do not
have to wait in real time for the test to complete.

* [Heartbeat] Use `maxFailedPings` var in `testCore`

Instead of asserting with the number literal equal to `maxFailedPings`,
use the variable `maxFailedPings` itself. Minor test fix.

* [Heartbeat] Fix a docstring in `testCommandsCore`

Fix the docstring of `contentFromMockContextSend` just so the mentioned
argument name matches the actual name of the argument of the said
function.

* [Heartbeat] Mock out background loop in fixture

Mock out `Core._loop` in `cogHeartbeat` fixture to avoid the background
loop being run. This allows us to say that the background loop does not
run upon cog instantiation via the fixture, rather than cancelling the
background task knowing that the background loop might have run for a
bit. The latter poses uncertainty and is not preferable.