[YOURLS] Bump `yourls-python` to 1.2.3 (#642)
This fixes an issue where the `[p]yourls stats` command would error
out. The root cause is in the call to YOURLS' `stats` method: when
the `limit` value is more than the number of shorten URLs, it would
raise a `KeyError`. This call is made by `cmdStats` (executed as part
of the `[p]yourls stats` command) which always specifies a `limit` of
3.