[Gatekeep] Add new cog (#666)
New cog to catch spam messages from first-time messages from recently joined accounts

* [Gatekeep] Initial commit of cog, should be ready to deploy

* Implemented changes properly, I got trolled by VSCode user interface

* Completed the remaining fixes for deployment

* [Gatekeep] Handled requests for 2025/02/09

* [Gatekeep] Fixed formatting for Tox style test

Lots of reformatting of the code to match the standards and conventions that the Tox style test. Hoping to not get a big fat red ❌ on my commit this time.

* [Gatekeep] Reformatted for Tox style test (again)

I swear to god these did not get flagged in the previous commit. Please do not give me the big fat red ❌ again