[Gatekeep] Restructure command tree, add new commands (#667)
* [Gatekeep] Restructured command tree, new commands

Commands relating to words and users have been moved into their respective subcommand trees. Added the ability to manually add and remove user IDs from the watchlist.

* [Gatekeep] Fixed bug with daily watch list check

Due to the way the watch list removed its own entries in the loop, it would skip every other ID in the list. This new fix addresses that issue.

* [Gatekeep] Updated ID arguments to user arguments

Commands for manually adding/removing users to the watch list now accept a discord.User object, which interprets both user IDs and usernames. Also modified pagination size for the word lists from 300 to 200. 300 characters was causing embeds to become too tall. Lastly, added a method of testing messages and evaluating scores.

* [Gatekeep] Tox style test fix

One line had to be fixed ICANT

* [Gatekeep] Fixed bug with repeat IDs being added

Forgot to change an 'id' to 'user.id', which cause the if statement to always evaluate to true. Fixed, now adding the same user multiple times does not work, they can only be added once.

* [Gatekeep] Removed unnecessary user.id checks

The user.id checks in addUser and removeUser were removed, as the discord.User class should handle invalid user IDs automatically. Also removed the Optional[str] in testMsg.

* [Gatekeep] Fixed bug with repeat IDs being added

Forgot to change an 'id' to 'user.id', which cause the if statement to always evaluate to true. Fixed, now adding the same user multiple times does not work, they can only be added once.

* [Gatekeep] Cleaned code with async and await

Reworked instances of accessing the watch list and word list to be a bit more efficient.

* [Gatekeep] Tox style

As per usual, there is only one offending line that needed to be changed in the entire script.